test(runtime-client): the tests move into a `test/` tree
`unit-test-coding-style.md` puts a package's unit tests in a `test`
directory that mirrors the source tree, and every one of this package's
thirteen sat beside its source instead. They now mirror: `test/`,
`test/backends/`, `test/client/`.

The move pays for itself beyond tidiness. The package's test task
globbed `*.test.ts backends/*.test.ts client/*.test.ts` -- a
hand-maintained list of directories, so a test file added under any
other subdirectory would have been silently skipped, running green
while testing nothing. The task is now `test/`, which recurses, and the
list has nothing left to fall out of date with.

Nothing but import specifiers changed inside the files: 45 rewritten so
each resolves to the same module it did before, and every changed line
in the thirteen is an import line. The run's inventory is identical --
321 lines of describe and it names, in the same order, before and after.
