+++ packages/runtime-client/deno.jsonc
@@ -3,7 +3,7 @@
   "name": "@commonfabric/runtime-client",
   "tasks": {
     // --no-check: this package is type-checked by `deno task check` (tasks/check.sh).
-    "test": "deno test --no-check --allow-env --allow-read --allow-ffi *.test.ts backends/*.test.ts client/*.test.ts",
+    "test": "deno test --no-check --allow-env --allow-read --allow-ffi test/",
     "integration": "LOG_LEVEL=warn deno test -A $INTEGRATION_TEST_FLAGS \"./integration/*.test.ts\""
   },
   "exports": {
