+++ packages/runtime-client/test/backends/web-worker-console-bridge.test.ts
@@ -5,15 +5,15 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 import { getLogger } from "@commonfabric/utils/logger";
-import { CompilerStackLoadError } from "../../runner/src/harness/deferred-compiler-stack.ts";
+import { CompilerStackLoadError } from "../../../runner/src/harness/deferred-compiler-stack.ts";
 import {
   ClientNotificationType,
   isWorkerConsoleNotification,
   RequestType,
   RuntimeErrorCode,
   TransportNotificationType,
-} from "../protocol/mod.ts";
-import { RuntimeProcessor } from "./mod.ts";
+} from "../../protocol/mod.ts";
+import { RuntimeProcessor } from "../../backends/mod.ts";
 
 // The worker entry (`backends/web-worker/index.ts`) installs a `message`
 // listener on `self` (which is `globalThis` under Deno) and reads
@@ -49,7 +49,7 @@ describe("web worker console bridge", () => {
     try {
       // Importing registers the message listener and posts the ready
       // notification.
-      await import("./web-worker/index.ts");
+      await import("../../backends/web-worker/index.ts");
       expect(posted).toContainEqual({
         type: TransportNotificationType.WorkerReady,
       });
@@ -246,7 +246,7 @@ describe("web worker request ledger and timing", () => {
       )) as typeof RuntimeProcessor.initialize;
 
     try {
-      await import("./web-worker/index.ts");
+      await import("../../backends/web-worker/index.ts");
       // Snapshot every counter this test asserts on: the loggers are global
       // singletons, so earlier tests in this file may already have ticked
       // some of them.
@@ -430,7 +430,7 @@ describe("web worker request ledger and timing", () => {
     // 100ms sample: the tick due during the block can only fire late, so a
     // positive workerLag is recorded. Counter existence only — the magnitude
     // is never asserted.
-    await import("./web-worker/index.ts");
+    await import("../../backends/web-worker/index.ts");
     const before = getLogger("runner.loop").getTimeStats("workerLag")?.count ??
       0;
     const end = performance.now() + 110;
